Tests/LibWeb/Text/expected/cross-origin-isolated-without-coop-coep.txt

crossOriginIsolated: false
SharedArrayBuffer serialization threw: DataCloneError
